AGDEV-11868 Removed/fixed old integration specs, fixed crashes during specs

+++ spec/integration/gdal/raster_band_algorithms_spec.rb
@@ -2,12 +2,13 @@ require 'spec_helper'
 require 'gdal/dataset'
 require 'gdal/raster_band'
 
-RSpec.describe GDAL::RasterBand do
+RSpec.describe GDAL::RasterBand, type: :integration do
   describe '#sieve_filter!' do
     let(:source_image_path) { 'spec/support/images/osgeo/geotiff/zi_imaging/image0.tif' }
     let(:dest_image_path) { 'tmp/image0.tif' }
+
     before do
-      FileUtils.rm(dest_image_path)
+      FileUtils.rm(dest_image_path) if File.exist?(dest_image_path)
       FileUtils.cp(source_image_path, dest_image_path)
     end
